pyproject.toml
[tool.poetry]
name = "practice-injector"
version = "0.1.0"
description = ""
authors = ["kunihik0 <tokuko2yuu.nikkorishiyou@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
injector = "^0.20.1"
pydantic = "^1.10.2"
mecab-python3 = "^1.0.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^0.991"
pytest = "^7.2.0"
pytest-xdist = "^3.1.0"
pytest-cov = "^4.0.0"
isort = "^5.11.4"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_first_party = ["src"]

[tool.black]
line-length = 88
skip-string-normalization = false
target-version = ['py39', 'py310']

[[tool.mypy.overrides]]
module = [
    "src.*"
]
ignore_missing_imports = true

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*_test.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@abstractmethod"
]
